+++ crypto_kem/smaugt3/m4/hwt.c
@@ -11,11 +11,11 @@
  * Arguments:   - uint8_t *res: pointer to ouptput polynomial r(x)
  *                (of length LWE), assumed to be already initialized
  *              - uint8_t *input: pointer to input seed (of length input_size)
- *              - size_t input_size: length of input seed
+ *              - unsigned long long input_size: length of input seed
  *              - uint16_t hmwt: hamming weight to sample
  **************************************************/
 void hwt(uint8_t *res, uint8_t *cnt_arr, const uint8_t *input,
-         const size_t input_size, const uint16_t hmwt) {
+         const unsigned long long input_size, const uint16_t hmwt) {
 
     uint32_t i = 0, pos = 0;
     uint32_t buf[SHAKE256_RATE * 2] = {0};
@@ -50,7 +50,7 @@ void hwt(uint8_t *res, uint8_t *cnt_arr, const uint8_t *input,
     if (pos != hmwt)
         fprintf(stderr, "hwt sampling error\n");
 
-    size_t cnt_arr_idx = 0;
+    unsigned long long cnt_arr_idx = 0;
     for (i = 0; i < DIMENSION; ++i) {
         cnt_arr_idx = ((i & 0x700) >> 8) & (-(res[i] & 0x01));
         cnt_arr[cnt_arr_idx] += (res[i] & 0x01);
