+++ crypto_kem/smaugt5/m4/hwt.h
@@ -35,6 +35,6 @@ typedef shake256incctx keccak_state;
 
 #define hwt SMAUG_NAMESPACE(hwt)
 void hwt(uint8_t *res, uint8_t *cnt_arr, const uint8_t *input,
-         const size_t input_size, const uint16_t hmwt);
+         const unsigned long long input_size, const uint16_t hmwt);
 
 #endif // SMAUG_HWT_H
