+++ crypto_kem/smaugt5/m4/verify.h
@@ -7,8 +7,8 @@
 #include "parameters.h"
 
 #define verify SMAUG_NAMESPACE(verify)
-int verify(const uint8_t *a, const uint8_t *b, size_t len);
+int verify(const uint8_t *a, const uint8_t *b, unsigned long long len);
 #define cmov SMAUG_NAMESPACE(cmov)
-void cmov(uint8_t *r, const uint8_t *x, size_t len, uint8_t b);
+void cmov(uint8_t *r, const uint8_t *x, unsigned long long len, uint8_t b);
 
 #endif
